requirements.txt
supabase-py==2.0.0
python-dotenv==1.0.0
pydantic>=2.0.0
requests==2.31.0
rich==13.7.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
aiohttp==3.9.1
flipside==2.0.7
schedule==1.2.1 